refactor: assume enabled state in core logic
This change splits the core logic of the plugin into separate class, so
that it can be easily enabled/disabled, so we do not have to check for
enabled state in sync functions all the time.

This also have an advantage of plugin "hot-reloadiing": we do not need
to restart Qt Creator if we want to reconnect to neovim, if the
connection failed for some reason.
